Populate topology configs via Chef Search
Search string is configurable w/ an attribute.

By default, search for nodes that have this recipe applied, and
are in the same environment.

Due to Cassandra being infrastructure agnostic, it uses the
more generic terms `DC` (datacenter) and `RAC` (rack) to
denote where a particular node lives.

Some manual notation is needed to translate that to AWS-speak
(i.e., region = datacenter, AZ = rack).

Lastly, there's probably a better way to handle the output,
esp. the YAML, but this works well enough.