Refactor opscenter recipe to actually work
Notifying a resource that's only defined if an attribute it true,
yet said attribute is false most of the time, doesn’t work.

Ensure the various OpsCenter-master-specific things
are defined as resources, but only created/etc. if said attribute
is true.

Create the OpsCenter clusters folder, so we have a place to
store our manually-generated cluster config.